The beginning of a "cookbook".
A cookbook is a "format", common in computer science books, that exposes
a specific topic in a practical manner, rather than have to dig deeply
into dense concepts.

Most of the samples are not documented (yet), since it was meant for
personal use/tests. Newer commits will follow the new format.